Move files to have everything in the same folder
