+++ src/Action/Type/Dropdown/LinkDropdownItemActionType.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Kreyu\Bundle\DataTableBundle\Action\Type\DropdownItemActionType;
+namespace Kreyu\Bundle\DataTableBundle\Action\Type\Dropdown;
 
 use Kreyu\Bundle\DataTableBundle\Action\Type\AbstractActionType;
 use Kreyu\Bundle\DataTableBundle\Action\Type\LinkActionType;
